+++ projects/packages/identity-crisis/changelog/fix-idc-tests-cleanup
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Tests: moved tests from a separate files to the main tests file.
